pyproject.toml
[tool.poetry]
name = "test-task-weather"
version = "0.1.0"
description = ""
authors = ["Valera K. <vk1337atliberty@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.35"
aiosqlite = "^0.20.0"
python-dotenv = "^1.0.1"
httpx = "^0.27.2"
openpyxl = "^3.1.5"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
